add packages and views all details

+++ src/main/java/org/example/srilanka_tour_and_travel/repository/DestinationRepository.java
@@ -3,5 +3,5 @@ package org.example.srilanka_tour_and_travel.repository;
 import org.example.srilanka_tour_and_travel.entity.Destination;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DestinationRepository extends JpaRepository<Destination,Integer> {
+public interface DestinationRepository extends JpaRepository<Destination,Long> {
 }
