+++ src/main/java/org/example/srilanka_tour_and_travel/repository/HotelRepository.java
@@ -3,5 +3,5 @@ package org.example.srilanka_tour_and_travel.repository;
 import org.example.srilanka_tour_and_travel.entity.Hotel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface HotelRepository extends JpaRepository<Hotel,Integer> {
+public interface HotelRepository extends JpaRepository<Hotel,Long> {
 }
