+++ src/main/java/org/example/srilanka_tour_and_travel/repository/PromotionRepository.java
@@ -3,5 +3,5 @@ package org.example.srilanka_tour_and_travel.repository;
 import org.example.srilanka_tour_and_travel.entity.Promotion;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PromotionRepository extends JpaRepository<Promotion,Integer> {
+public interface PromotionRepository extends JpaRepository<Promotion,Long> {
 }
